fix: 데이터 load 오류 수정
일부 페이지에서 id값이 갱신 되었을 때 데이터를 다시 불러오지 못하는 문제 수정

+++ src/main/frontend/src/Routes/Board/hooks/useDeleteBoard.js
@@ -9,7 +9,7 @@ export const useDeleteBoard = () => {
   const { id } = useParams();
 
   const deleteBoard = () => {
-    id &&
+    if (window.confirm("정말로 삭제하겠습니까?")) {
       dispatch(deleteBoardById(id))
         .unwrap()
         .then((res) => {
@@ -17,6 +17,7 @@ export const useDeleteBoard = () => {
           toast.success("게시물 삭제 성공");
         })
         .catch((err) => toast.error("게시물 삭제 실패"));
+    }
   };
 
   return { deleteBoard };
